requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib
python-jose[cryptography]
python-multipart
pydantic-settings
alembic
greenlet
pydantic[email]
bcrypt
python-socketio
websockets
redis
pytest
PyJWT==2.9.0